Fixed some minor problems. Updated test project to netcore6.0
