+++ DotnetToolset.Patterns/Dddd/Implementations/Rules/IsNotNullRule.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using DotnetToolset.Patterns.Dddd.Enums;
 using Res = DotnetToolset.Patterns.Resources.Literals;
+using DotnetToolset.ExtensionMethods;
 
 namespace DotnetToolset.Patterns.Dddd.Implementations.Rules
 {
@@ -33,7 +34,7 @@ namespace DotnetToolset.Patterns.Dddd.Implementations.Rules
 				return (true, null);
 			}
 
-			return (false, Res.p_RuleIsNotNullNotPassed);
+			return (false, Res.p_RuleIsNotNullNotPassed.ParseParameters(new[] { value.ToString(), Rule.Value }));
 		}
 	}
 }
