+++ DotnetToolset.Patterns/DotnetToolset.Patterns.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <TargetFramework>netstandard2.1</TargetFramework>
 	  <RootNamespace>DotnetToolset.Patterns</RootNamespace>
-	  <Version>0.14.0</Version>
+	  <Version>0.14.1</Version>
 	  <Authors>Ahead Labs, S.L.</Authors>
 	  <Description>Opinionated miscellaneous patterns for multilayered .NET projects.</Description>
 	  <Copyright>Ahead Labs, S.L.</Copyright>
